requirements.txt
selenium
webdriver-manager
beautifulsoup4
openai
python-dotenv